+++ lib/src/model/message_context.dart
@@ -3,6 +3,7 @@ import 'package:chatterbox/src/model/sticker_model.dart';
 class MessageContext {
   final int userId;
   final int chatId;
+  final String? text;
   final int? editMessageId;
   final String? username;
   final StickerModel? sticker;
@@ -12,6 +13,7 @@ class MessageContext {
   MessageContext({
     required this.userId,
     required this.chatId,
+    this.text,
     this.editMessageId,
     this.username,
     this.sticker,