+++ lib/src/utils/update_processing_utils.dart
@@ -34,6 +34,7 @@ void processTextInternal(Message message, UpdateMessageHandler commandHandler) {
     chatId: chatId,
     editMessageId: message.messageId,
     username: user?.username,
+    text: message.text,
   ));
 }