+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.4
+- Add Message text inside MessageContext instead of appending to args
+- Add args field to toStepUri util function
+
 ## 1.0.3
 - Fix working with pending message